# docs/04-api/openapi/frameflow-v1.yaml
openapi: 3.0.3
info:
  title: FrameFlow Select API
  version: 1.0.0
  description: FrameFlow Select 批量质检与 Top-K 优选平台。身份端点沿用 M01（/api/v1/auth、/api/v1/teams）；产品端点为手写 OpenAPI 权威契约，运行时以 springdoc 输出并经验证对齐。
...
paths:
...
  /api/v1/quality-profile-versions/{versionId}/publish:
    post:
      tags:
      - product
      summary: 发布模板版本（不可变，仅 OWNER）
      operationId: publishProfileVersion
      parameters:
      - name: versionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/QualityProfileVersion'
      security:
      - bearerAuth: []
  /api/v1/projects:
    get:
      tags:
      - product
      summary: 团队项目列表
      operationId: listProjects
      parameters:
      - name: X-Team-Id
        in: header
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
      security:
      - bearerAuth: []
    post:
      tags:
      - product
      summary: 创建项目（OWNER/OPERATOR）
      operationId: createProject
      parameters:
      - name: X-Team-Id
        in: header
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
        required: true
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: integer
                  format: int64
      security:
      - bearerAuth: []
  /api/v1/projects/{projectId}/quality-profiles:
    get:
      tags:
      - product
      summary: 项目质量模板列表
      operationId: listQualityProfiles
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QualityProfile'
      security:
      - bearerAuth: []
    post:
      tags:
      - product
      summary: 创建质量模板草稿（OWNER/OPERATOR）
      operationId: createQualityProfile
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
        required: true
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: integer
                  format: int64
      security:
      - bearerAuth: []
  /api/v1/projects/{projectId}/quality-profiles/{profileId}/versions:
    post:
      tags:
      - product
      summary: 保存质量模板版本草稿
      operationId: createProfileVersion
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: profileId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileVersionRequest'
        required: true
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/QualityProfileVersion'
      security:
      - bearerAuth: []
...
  /api/v1/projects/{projectId}:
    get:
      tags:
      - product
      summary: 获取项目
      operationId: getProject
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Project'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    QualityProfileVersion:
      type: object
      properties:
        id:
          type: integer
          format: int64
        profileId:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        status:
          type: string
        payload:
          type: string
        payloadDigest:
          type: string
        publishedAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    CreateProfileRequest:
      type: object
      properties:
        name:
          type: string
        templateType:
          type: string
    CreateProfileVersionRequest:
      type: object
      properties:
        payload:
          type: string
...
    Project:
      type: object
      properties:
        id:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        status:
          type: string
        archivedAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QualityProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        name:
          type: string
        templateType:
          type: string
        status:
          type: string
        createdBy:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  headers:
    RequestId:
      description: 服务器为本次 HTTP 请求生成的请求 ID
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT